broke up main
